ensuremembers: only check presence ext. accounts
The previous version of yensuremembers verified that all users
accounts in the user list are present on the system. However, this
is not really needed when working with internal accounts, since
they automatically get created if they are not present.

Add -i commandline option for specifying internal domains, and
only check that external user accounts (accounts not in internal
domains) are present.