YDA-5885: add -H option to group lifecycle report
Add option for displaying sizes in human-readable format to group
lifecycle report, and change default behaviour to not display
sizes in human-readable format.

Not reporting in human-readable format makes it easier to perform
calculations on size data in the reports, e.g. by adding up size
data for selected groups in a spreadsheet.